requirements.txt
flask
jwt
python-dotenv
requests
